referenceField: properly handle default props
Currently, the referencefield label is only shown when
addLabel is set. But by default, the label field should
show the label (based on default value).

So now the referenceField now shown the label by default
whithout use addLabel props.

This bug is due to the use of redux who ignored the
default props of the component.